+++ examples/restful/webspicy/config.rb
@@ -1,0 +1,1 @@
+Webspicy::Configuration.new(Path.dir)