+++ examples/restful/webspicy/rack.rb
@@ -1,0 +1,6 @@
+Webspicy::Configuration.inherits(Path.dir) do |c|
+  c.client = Webspicy::RackTestClient.for(::Sinatra::Application)
+  c.before_each do |_,_,_,client|
+    client.api.post "/reset"
+  end
+end