+++ spec/unit/configuration/config.rb
@@ -1,0 +1,3 @@
+Webspicy::Configuration.new(Path.dir) do |c|
+  c.precondition Class.new
+end